[lx]修改更多需求

+++ BeautyHouse_iphone/ViewControllers/MyOrder/2.0新增/ScrollPageView.m
@@ -290,7 +290,7 @@
     //[self setupRefresh:@"current"];
     CustomTableView *vTableContentView =(CustomTableView *)[_contentItems objectAtIndex:_mCurrentPage];
     
-    [vTableContentView setupRefresh:[NSString stringWithFormat:@"%li",_mCurrentPage]];
+    [vTableContentView setupRefresh:[NSString stringWithFormat:@"%li",(long)_mCurrentPage]];
